Modify 'AddBtn' listener to module
