+++ js/index.js
@@ -1,6 +1,7 @@
 import {bookList} from './modules/books.mjs'
 import {fromLS} from './modules/localStorage.mjs'
+import {displayBook} from './modules/displayBook.mjs'
+import {addBtn} from './modules/addbook.mjs'
 
 fromLS()
-
-console.log(bookList);
+displayBook()
