+++ js/modules/elements.mjs
@@ -2,5 +2,7 @@
 const navlist = document.querySelectorAll('.navlink');
 const section = document.querySelectorAll('.section');
 const bookSection = document.getElementById('bookList');
+const newBook = document.getElementById('add');
+const newBookEl = newBook.querySelectorAll('input');
 
-export {navlist, section, bookSection}
+export {navlist, section, bookSection, newBook, newBookEl}
