+++ script.js
@@ -45,71 +45,71 @@
 // // Create book card for each e in bookList
 // const bookSection = document.getElementById('bookList');
 
-function clearAll() {
-  const remove = bookSection.querySelectorAll('.book');
-  remove.forEach((book) => {
-    bookSection.removeChild(book);
-  });
-}
+// function clearAll() {
+//   const remove = bookSection.querySelectorAll('.book');
+//   remove.forEach((book) => {
+//     bookSection.removeChild(book);
+//   });
+// }
+//
+// function displayBook() {
+//   clearAll();
+//
+//   let count = 1;
+//
+//   bookList.list.forEach((book) => {
+//     const bDiv = document.createElement('div');
+//     const nameDiv = document.createElement('div');
+//     const bTitle = document.createElement('p');
+//     const bAuthor = document.createElement('small');
+//     const bBtn = document.createElement('button');
+//
+//     bDiv.classList.add('book');
+//     bTitle.innerHTML = book.title;
+//     bAuthor.innerHTML = book.author;
+//     bBtn.innerHTML = 'Remove';
+//
+//     // Background flipflop
+//     count += 1;
+//     if (count % 2 !== 0) {
+//       bDiv.classList.add('gray');
+//       count = 1;
+//     }
+//
+//     nameDiv.appendChild(bTitle);
+//     nameDiv.appendChild(bAuthor);
+//     bDiv.appendChild(nameDiv);
+//     bDiv.appendChild(bBtn);
+//
+//     bookSection.appendChild(bDiv);
+//
+//     // Remove listener
+//     bBtn.addEventListener('click', () => {
+//       bookSection.removeChild(bDiv);
+//       return bookList.remove(book.title, book.author);
+//     });
+//   });
+// }
 
-function displayBook() {
-  clearAll();
+// displayBook();
 
-  let count = 1;
+// // html Addbutton function
+// const newBook = document.getElementById('add');
+// const newBookEl = newBook.querySelectorAll('input');
 
-  bookList.list.forEach((book) => {
-    const bDiv = document.createElement('div');
-    const nameDiv = document.createElement('div');
-    const bTitle = document.createElement('p');
-    const bAuthor = document.createElement('small');
-    const bBtn = document.createElement('button');
-
-    bDiv.classList.add('book');
-    bTitle.innerHTML = book.title;
-    bAuthor.innerHTML = book.author;
-    bBtn.innerHTML = 'Remove';
-
-    // Background flipflop
-    count += 1;
-    if (count % 2 !== 0) {
-      bDiv.classList.add('gray');
-      count = 1;
-    }
-
-    nameDiv.appendChild(bTitle);
-    nameDiv.appendChild(bAuthor);
-    bDiv.appendChild(nameDiv);
-    bDiv.appendChild(bBtn);
-
-    bookSection.appendChild(bDiv);
-
-    // Remove listener
-    bBtn.addEventListener('click', () => {
-      bookSection.removeChild(bDiv);
-      return bookList.remove(book.title, book.author);
-    });
-  });
-}
-
-displayBook();
-
-// html Addbutton function
-const newBook = document.getElementById('add');
-const newBookEl = newBook.querySelectorAll('input');
-
-newBookEl[2].addEventListener('click', () => {
-  if (newBookEl[0].value !== '' && newBookEl[1].value !== '') {
-    bookList.add(`"${newBookEl[0].value}"`, `by ${newBookEl[1].value}`);
-    displayBook();
-    newBookEl[0].value = '';
-    newBookEl[1].value = '';
-  } else if (newBookEl[0].value !== '') {
-    bookList.add(newBookEl[0].value, 'Anonymous');
-    displayBook();
-    newBookEl[0].value = '';
-    newBookEl[1].value = '';
-  }
-});
+// newBookEl[2].addEventListener('click', () => {
+//   if (newBookEl[0].value !== '' && newBookEl[1].value !== '') {
+//     bookList.add(`"${newBookEl[0].value}"`, `by ${newBookEl[1].value}`);
+//     displayBook();
+//     newBookEl[0].value = '';
+//     newBookEl[1].value = '';
+//   } else if (newBookEl[0].value !== '') {
+//     bookList.add(newBookEl[0].value, 'Anonymous');
+//     displayBook();
+//     newBookEl[0].value = '';
+//     newBookEl[1].value = '';
+//   }
+// });
 
 // Date
 const datePlace = document.querySelector('.date');
